backend/requirements.txt

fastapi>=0.108.0
uvicorn>=0.25.0
pydantic>=2.6.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
sqlalchemy>=2.0.25
alembic>=1.13.1
pymongo>=4.6.1
boto3>=1.34.19
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
